Move filters button in separate helper
