+++ app/helpers/dorsale/all_helpers.rb
@@ -3,6 +3,7 @@ module Dorsale
     include ButtonHelper
     include CommentsHelper
     include ContextHelper
+    include FiltersHelper
     include FormHelper
     include LinkHelper
     include PaginationHelper
